pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "phi_combo"
version = "0.1.0"
description = "Hugging Face dataset downloader with YAML configuration"
requires-python = ">=3.8"
dependencies = [
    "datasets>=2.0.0",
    "pyyaml>=6.0",
    "requests>=2.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.pyright]
extraPaths = ["."]
